syscall_intercept: Fix instruction operand decoding in disasm_wrapper.c
The operand decoding loop was removed in a commit, expect for the case of
a jump instruction. But the loop is needed also to find out about other
non-relocatable instructions, e.g.:

mov %rax, 0x36eb55d(%rip)

In the case of the instruction above, the loop is needed to recognize that
the second operand relies on the value of the RIP register, thus can not be
relocated.